feat: change url api lowercase

+++ Ecommerce.api/Service/ProductService.cs
@@ -1,3 +1,4 @@
+using System.ComponentModel.DataAnnotations;
 using AutoMapper;
 using AutoMapper.QueryableExtensions;
 using Ecommerce.api.Dto;
@@ -98,6 +99,8 @@ public class ProductService(IProductRepository productRepository,IMapper mapper)
 
     public async Task  DeleteAsync(int id)
     {
+        if (id == 0)
+            throw new ValidationException(nameof(id) + " cannot be 0");
         await productRepository.DeleteAsync(id);
        
     }
